efficacious `#[must_use]` methods for nightly, fix warnings
137660d9 annotated many methods with `#[must_use]`, but despite not
being rejected by the compiler, this is actually a no-op (as of Rust
1.20). `#[must_use]` for functions is now available on nightly
Rust (since rust-lang/rust@3645b062), under a feature-gate (since
rust-lang/rust@8492ad24 and 35c44941), with a warning if `#[must_use]`
appears on a function or method without the feature being enabled—a
warning that regrettably added a lot of noise to Euclid's build, as
reported in rust-lang/rust#44213. But with judicious use of `cfg_attr`,
Euclid can make use of this functionality if our "unstable" feature is
enabled, and shut up the warnings in any case.