transform3d: Introduce Transform3D::project_to_2d to flatten a transform.
Need to use this for https://github.com/servo/webrender/issues/3394.

The normalization of perspective is not necessary but Gecko does it as a
performance optimization, and ScaleOffset in WR wouldn't handle this.

See: https://bugzilla.mozilla.org/show_bug.cgi?id=1241161

But in any case I'm happy to teach ScaleOffset about that case instead if you
think it's better.