Allow building non-`Trig` transforms that don't rotate
The `Trig` trait bound is only required for doing rotations, and also `Trig` is
only implemented for `f64` and `f32`. Putting the `Trig` trait bound on the main
`impl` for `Transform2D` methods was a pretty restrictive and applied
restrictions more often than was needed.

This commit moves the `Trig` trait bound to an `impl` specifically for
rotation-related methods of `Transform2D`, and the result is that we can now
scale or translate things represented with `i32` or `i64` or other non-`Trig`
scalars.