Implement `distance_to` method for points
I don't know, if such a simple helper method is considered acceptable
for this crate, but I figured the best way of asking that question is by
just submitting a pull request.

I'm also not sure about the return value. It seems like returning
`Length` would be the right thing to do, but that's not what the
`length` methods return. I figured I'd follow that established
convention, but I'm happy to make changes as requested.