Add `Vector2D::try_normalize` and `Vector3D::try_normalize`.
These methods check that the length of the vector is not zero and return None in
that case.  It is slightly cleaner than either checking that the vector is
not *equal* to zero or checking that the length is not zero before calling the
regular `normalize` method.

If the vector has very small component values, length may be calculated as zero,
and thus the length of the normalized vector will be infinity.