Fix compile of `arbitrary` feature.
`derive(arbitrary::Arbitrary)` uses paths from `std`, which is
incompatible with `#![no_std]`. The simplest path forward is to use a
manual implementation instead of the derive.

An alternative to this strategy would be to change the conditional
`no_std` to be `not(any(test, feature = "arbitrary")`, but this seems
simpler and more robust.